# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/company_tax_settingses:
    get:
      tags:
        - CompanyTaxSettingses
      summary: List CompanyTaxSettingses
      description: Get a paginated list of CompanyTaxSettingses with optional filtering and sorting
      operationId: listCompanyTaxSettingses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyTaxSettingsList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CompanyTaxSettingses
      summary: Create CompanyTaxSettings
      description: Create a new CompanyTaxSettings
      operationId: createCompanyTaxSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyTaxSettingsRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyTaxSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/e_faktur_documents:
    get:
      tags:
        - EFakturDocuments
      summary: List EFakturDocuments
      description: Get a paginated list of EFakturDocuments with optional filtering and sorting
      operationId: listEFakturDocuments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EFakturDocumentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EFakturDocuments
      summary: Create EFakturDocument
      description: Create a new EFakturDocument
      operationId: createEFakturDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEFakturDocumentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EFakturDocument'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    CompanyTaxSettings:
      type: object
      required:
        - id
        - rounding_method
        - default_exigibility
        - metadata
      properties:
        id:
          type: string
          format: uuid
        rounding_method:
          type: string
        default_exigibility:
          type: string
        cash_basis_transition_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCompanyTaxSettingsRequest:
      type: object
      required:
        - rounding_method
        - default_exigibility
        - metadata
      properties:
        rounding_method:
          type: string
        default_exigibility:
          type: string
        cash_basis_transition_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CompanyTaxSettingsList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompanyTaxSettings'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    EFakturDocument:
      type: object
      required:
        - id
        - tax_transaction_id
        - number
        - transaction_code
        - taxpayer_segment
        - period
        - sequence
        - assignment_date
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        tax_transaction_id:
          type: string
          format: uuid
        number:
          type: string
          maxLength: 20
        transaction_code:
          type: string
          maxLength: 3
        taxpayer_segment:
          type: string
          maxLength: 3
        period:
          type: string
          format: date
        sequence:
          type: integer
          format: int32
        assignment_date:
          type: string
          format: date
        status:
          type: string
        replaces_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateEFakturDocumentRequest:
      type: object
      required:
        - tax_transaction_id
        - number
        - transaction_code
        - taxpayer_segment
        - period
        - sequence
        - assignment_date
        - status
        - metadata
      properties:
        tax_transaction_id:
          type: string
          format: uuid
        number:
          type: string
          maxLength: 20
        transaction_code:
          type: string
          maxLength: 3
        taxpayer_segment:
          type: string
          maxLength: 3
        period:
          type: string
          format: date
        sequence:
          type: integer
          format: int32
        assignment_date:
          type: string
          format: date
        status:
          type: string
        replaces_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    EFakturDocumentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EFakturDocument'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []